Added Basic Player Movement
I've created a new scene named "Level 01" which contains a Player character and a basic environment. The player character was assigned a new C# script  called "PlayerMovement" which allows the player to move with WASD or the arrow keys. It also uses the mouse position to determine where the player is facing.  The Input settings found in Edit - Project Settings - Input were also modified for easier use.