Modified the Main Camera
I modified the Main Camera to use Orthographic projection instead of Perspective because this suited the Isometric Camera style that we had chosen for the game.  A C# script called "CameraFollow" was created and assigned to the main camera. This script makes the main camera smoothly follow the Player around the scene. I also noticed that the Player would no longer be visible after walking behind certain walls so created a C# script called "TransparentWalls" which made the walls become transparent as the Player got close to them so that the Player would still be visible.