Added Player Model to Game with Animations
I added Leon's Player Model to the game. I used a website called Mixamo to auto rig the model and get suitable animations. I set up an Animator Controller in Unity and made some changes to the scripts to get the animations to work. Unfortunately there was an issue with the materials of the player so this will need to be fixed at a later date.