Got Leon's Level working in the Game and other changes.
I got Leon's level mostly working in the game but some objects in the room were edited. I organised the level into  Walls and Floors. I also tried to get the transparent walls working again but it made the game really slow so I disabled it. I also made some changes to the player movement to fix a glitch where if the player held the mouse over the character while moving the character wouldn't know where to go so the animations would go weird. I also made some changes to the camera.